[kv_offload] Merge FilterReusedOffloadingManager into cpu/manager.py and add OffloadingManager.request_finished
This commit implements Tasks 5 and 6 of #33689:

- Task 5: Move `reuse_manager.py` into `cpu/manager.py`
- Task 6: Add `request_finished` lifecycle hook to `OffloadingManager`
